backend/pyproject.toml
[project]
name = "travel-agent-backend"
version = "0.1.0"
description = "Clarification-first travel planning API"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.34.0",
  "pydantic>=2.10.0",
  "python-dotenv>=1.0.1"
]

[tool.uv]
package = false
